Unsigned handling of quads:  - Longs are now used where appropriate - note: other quads are not changed as their sign is not relevant.

+++ src/main/java/software/coley/llzip/part/ZipPart.java
@@ -12,7 +12,7 @@ public interface ZipPart {
 	/**
 	 * @return Length of current content.
 	 */
-	int length();
+	long length();
 
 	/**
 	 * @return Implementation type.
